pyproject.toml
[project]
name = "dog"
version = "0.1.0"
description = "Local RAG whatever"
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "faiss-cpu>=1.12.0",
    "ffmpeg-python>=0.2.0",
    "groq>=0.31.0",
    "numpy>=2.2.6",
    "ocrmypdf>=16.10.4",
    "pymupdf>=1.26.3",
    "sentence-transformers>=5.1.0",
    "torch>=2.8.0",
    "transformers>=4.55.2",
    "openai-whisper @ git+https://github.com/openai/whisper.git",
    "pytesseract>=0.3.13",
    "pillow>=11.3.0",
    "langchain-text-splitters>=0.3.9",
]
